+++ src/cgame/cg_main.cpp
@@ -515,6 +515,7 @@ vmCvar_t etj_autoLoad;
 vmCvar_t etj_uphillSteps;
 vmCvar_t etj_quickFollow;
 vmCvar_t etj_chatLineWidth;
+vmCvar_t etj_enableOB;
 
 typedef struct
 {
@@ -872,6 +873,7 @@ cvarTable_t cvarTable[] =
 	{ &etj_uphillSteps, "etj_uphillSteps", "1", CVAR_ARCHIVE },
 	{ &etj_quickFollow, "etj_quickFollow", "1", CVAR_ARCHIVE },
 	{ &etj_chatLineWidth, "etj_chatLineWidth", "62", CVAR_ARCHIVE },
+	{ &etj_enableOB, "etj_enableOB", "1", CVAR_ARCHIVE },
 };
 
 
@@ -982,7 +984,7 @@ void CG_UpdateCvars(void)
 					cv->vmCvar == &cg_hideMe || cv->vmCvar == &cg_noclipScale ||
 					cv->vmCvar == &etj_enableTimeruns || cv->vmCvar == &etj_noActivateLean ||
 					cv->vmCvar == &etj_touchPickupWeapons || cv->vmCvar == &etj_autoLoad ||
-					cv->vmCvar == &etj_quickFollow
+					cv->vmCvar == &etj_quickFollow || cv->vmCvar == &etj_enableOB
 				    )
 				{
 					fSetFlags = qtrue;
@@ -1078,7 +1080,8 @@ void CG_setClientFlags(void)
 									 ((etj_enableTimeruns.integer > 0) ? CGF_ENABLE_TIMERUNS : 0) |
 									 ((etj_noActivateLean.integer > 0) ? CGF_NOACTIVATELEAN : 0) |
 									 ((etj_autoLoad.integer > 0) ? CGF_AUTO_LOAD : 0) |
-									 ((etj_quickFollow.integer > 0) ? CGF_QUICK_FOLLOW : 0)
+									 ((etj_quickFollow.integer > 0) ? CGF_QUICK_FOLLOW : 0) |
+									 ((etj_enableOB.integer > 0) ? CGF_ENABLE_OB : 0)
 	                                 // Add more in here, as needed
 	                             ),