+++ src/game/bg_pmove.cpp
@@ -2332,7 +2332,11 @@ static void PM_GroundTrace(void)
 		PM_CrashLand();
 
 		// Disable overbounce?
-		// This breaks skim ramps when OB is disabled... Is it really required?
+		if (!pm->enableOB)
+		{
+			PM_ClipVelocity(pm->ps->velocity, pml.groundTrace.plane.normal, pm->ps->velocity, OVERCLIP);
+		}
+
 		if (pm->shared & BG_LEVEL_NO_OVERBOUNCE)
 		{
 			if (!((trace.surfaceFlags & SURF_OVERBOUNCE) != 0))
@@ -2359,6 +2363,19 @@ static void PM_GroundTrace(void)
 	pm->ps->groundEntityNum = trace.entityNum;
 
 	// Disable overbounce?
+	if (!pm->enableOB)
+	{
+		if (trace.plane.normal[2] == 1)
+		{
+			pm->ps->velocity[2] = 0;
+			if (trace.plane.type == 2 && pm->waterlevel < 2)
+			{
+				auto offset = pm->ps->origin[2] - trace.endpos[2];
+				pm->ps->origin[2] -= offset;
+			}
+		}
+	}
+
 	if (pm->shared & BG_LEVEL_NO_OVERBOUNCE)
 	{
 		if (trace.plane.normal[2] == 1 && !((trace.surfaceFlags & SURF_OVERBOUNCE) != 0))