+++ src/game/g_client.cpp
@@ -1641,6 +1641,7 @@ void ClientUserinfoChanged(int clientNum)
 	client->pers.noActivateLean = (client->pers.clientFlags & CGF_NOACTIVATELEAN) != 0 ? qtrue : qfalse;
 	client->pers.autoLoad = (client->pers.clientFlags & CGF_AUTO_LOAD) != 0 ? qtrue : qfalse;
 	client->pers.quickFollow = (client->pers.clientFlags & CGF_QUICK_FOLLOW) != 0 ? qtrue : qfalse;
+	client->pers.enableOB = (client->pers.clientFlags & CGF_ENABLE_OB) != 0 ? qtrue : qfalse;
 
 	// set name
 	Q_strncpyz(oldname, client->pers.netname, sizeof(oldname));