+++ src/game/g_local.h
@@ -835,6 +835,7 @@ typedef struct
 	qboolean noActivateLean;
 	qboolean autoLoad;
 	qboolean quickFollow;
+	qboolean enableOB;
 
 	unsigned int maxFPS;
 	char netname[MAX_NETNAME];